Modify Vector property validation to get around vectormath coercing shape
The current behaviour of vectormath creates lists of vectors, so with
input of shape (3,) it becomes shape (1, 3). This conflicts with the array
shape. To get around this temporarily, Vector properties override validate.
Once vector math separates vectors and lists of vectors, this Vectors
can return to using super.validate

See Issue 3ptscience/vectormath#7