Keep track of inheritance to enable translation
As the metaclass constructs HasProperties subclasses, it enters the baked-in
inheritance structure into a dictionary that is stored on the class.
This makes it possible to re-construct the data object inheritance path
in its original form. This is useful for displaying the class structure
(i.e., UML graphs), but also for replicating / reproducing the inheritance
when writing translation libraries for data objects.